Conflicts:
	Projects/UnitTests/Classes/DeviceInfoTest.cpp
	Projects/UnitTests/TemplateProjectiPhone.xcodeproj/project.pbxproj
	Projects/iPhoneFramework/Framework.xcodeproj/project.pbxproj
	Sources/Internal/Platform/TemplateAndroid/DeviceInfoAndroid.cpp
	Sources/Internal/Platform/TemplateAndroid/DeviceInfoAndroid.h
	Sources/Internal/UI/UIControl.cpp
	Sources/Internal/UI/UIControl.h
	Sources/Internal/UI/UIStaticText.cpp
	Tools/ResourcePacker/Classes/FrameworkMain.cpp (reverse-merged from commit b18a1f8afc6ffc07735179cc473d6188d59fdbcb)

+++ Sources/Internal/Render/Highlevel/Vegetation/RenderBatchPool.cpp
@@ -90,7 +90,7 @@ RenderBatch* RenderBatchPool::Get(NMaterial* key, VegetationMaterialTransformer*
     RenderBatch* rb = NULL;
     
     size_t currentPoolSize = entry->renderBatches.size();
-    if(currentPoolSize <= (size_t)entry->poolLine)
+    if(currentPoolSize <= entry->poolLine)
     {
         rb = CreateRenderBatch(key, entry, transform);
     }
